Statistical Analysis Pandas
Statistical Analysis. Pandas
One of the main goals of classification machine learning models is to try to find patterns that will allow the algorithm to separate the classes of samples. Usually, they do that using different approaches. People can do that too, and as a data scientist, it is important to find manually the patterns that will allow a Machine Learning algorithm to do that. Usually, we use that Statistical tools and methods.