Vector form of the equation
The data in this form is quite huge, and we can’t perform operations on them one by one. So for this, vector operations come to the rescue, and we, in machine learning, use vector operations.

So the vector form of the above equation can be written as:

Y = Theta.(dot)X

The NumPy dot product is used. NumPy gives us a lot of functions for performing mathematical matrix operations.

To get the solution, we have to transpose the theta, i.e., replace rows with columns and then dot the product with the feature matrix. This is done because of the multiplicative properties of matrices which state that — the number of rows of the first matrix should be equal to the number of columns of the second matrix.